+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "comint" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Fix password prompt bug where it would show multiple times the first time a password prompt is encountered, 
+  and then never work again after the first time. It was a regexp issue.
+
 ## [0.0.2]
 
 - Initial release. Highly experimental.